+++ Assets/Scripts/Assembly-CSharp/LeaderboardDialog.cs
@@ -563,7 +563,7 @@ public class LeaderboardDialog : TextDialog
 
 	private int SeasonDaysLeft()
 	{
-		return (0 - Singleton<TimeManager>.Instance.ServerTime.DayOfWeek + 7) % 7;
+		return (0 - (int)Singleton<TimeManager>.Instance.ServerTime.DayOfWeek + 7) % 7;
 	}
 
 	public void OpenCupView()
